[Fix] pass data tp node (int)

+++ 0x17-doubly_linked_lists/2-add_dnodeint.c
@@ -23,7 +23,7 @@ dlistint_t *add_dnodeint(dlistint_t **head, const int n)
 	if (!new)
 		return (NULL);
 
-	new->n = n;
+	new->n = (int) n;
 	new->prev = NULL;
 	new->next = (*head) ? *head : NULL;
 	*head = new;
